Evictor interface refactored according to review comments
