+++ hazelcast/src/main/java/com/hazelcast/map/impl/eviction/EvictorImpl.java
@@ -19,16 +19,25 @@ package com.hazelcast.map.impl.eviction;
 import com.hazelcast.config.EvictionPolicy;
 import com.hazelcast.config.MapConfig;
 import com.hazelcast.config.MaxSizeConfig;
+import com.hazelcast.map.impl.MapContainer;
 import com.hazelcast.map.impl.MapServiceContext;
+import com.hazelcast.map.impl.PartitionContainer;
 import com.hazelcast.map.impl.record.Record;
 import com.hazelcast.map.impl.recordstore.RecordStore;
+import com.hazelcast.nio.Address;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.partition.InternalPartition;
 import com.hazelcast.partition.InternalPartitionService;
+import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.util.Clock;
+import com.hazelcast.util.MemoryInfoAccessor;
+import com.hazelcast.util.RuntimeMemoryInfoAccessor;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.Iterator;
+import java.util.List;
 
 import static com.hazelcast.cluster.memberselector.MemberSelectors.DATA_MEMBER_SELECTOR;
 
@@ -38,40 +47,54 @@ import static com.hazelcast.cluster.memberselector.MemberSelectors.DATA_MEMBER_S
 public class EvictorImpl implements Evictor {
 
     protected static final int ONE_HUNDRED_PERCENT = 100;
+    protected static final int EVICTION_START_THRESHOLD_PERCENTAGE = 95;
+    protected static final int ONE_KILOBYTE = 1024;
+    protected static final int ONE_MEGABYTE = ONE_KILOBYTE * ONE_KILOBYTE;
 
     protected final MapServiceContext mapServiceContext;
-    protected final EvictionChecker evictionChecker;
 
-    public EvictorImpl(EvictionChecker evictionChecker, MapServiceContext mapServiceContext) {
-        this.evictionChecker = evictionChecker;
+    // not final for testing purposes.
+    protected MemoryInfoAccessor memoryInfoAccessor;
+
+    public EvictorImpl(MapServiceContext mapServiceContext) {
         this.mapServiceContext = mapServiceContext;
+        this.memoryInfoAccessor = new RuntimeMemoryInfoAccessor();
     }
 
     @Override
-    public EvictionChecker getEvictionChecker() {
-        return evictionChecker;
+    public void evict(RecordStore recordStore) {
+        if (!recordStore.isEvictionEnabled()
+                || !isReachedMaxSize(recordStore)) {
+            return;
+        }
+
+        int removalSize = calculateRemovalSize(recordStore);
+        if (removalSize < 1) {
+            return;
+        }
+
+        removeRecords(recordStore, removalSize);
     }
 
-    @Override
-    public void removeSize(int removalSize, RecordStore recordStore) {
+    protected void removeRecords(RecordStore recordStore, int sizeToEvict) {
         long now = Clock.currentTimeMillis();
         MapConfig mapConfig = recordStore.getMapContainer().getMapConfig();
 
         boolean backup = isBackup(recordStore);
 
-        final EvictionPolicy evictionPolicy = mapConfig.getEvictionPolicy();
+        EvictionPolicy evictionPolicy = mapConfig.getEvictionPolicy();
         // criteria is a long value, like last access times or hits,
         // used for calculating LFU or LRU.
-        final long[] criterias = createAndPopulateEvictionCriteriaArray(recordStore, evictionPolicy);
+        long[] criterias = createAndPopulateEvictionCriteriaArray(recordStore, evictionPolicy);
         if (criterias == null) {
             return;
         }
         Arrays.sort(criterias);
         // check in case record store size may be smaller than evictable size.
-        final int evictableBaseIndex = getEvictionStartIndex(criterias, removalSize);
-        final long criteriaValue = criterias[evictableBaseIndex];
+        int evictableBaseIndex = getEvictionStartIndex(criterias, sizeToEvict);
+        long criteriaValue = criterias[evictableBaseIndex];
         int evictedRecordCounter = 0;
-        final Iterator<Record> iterator = recordStore.iterator();
+        Iterator<Record> iterator = recordStore.iterator();
         while (iterator.hasNext()) {
             Record record = iterator.next();
             Data key = record.getKey();
@@ -81,7 +104,7 @@ public class EvictorImpl implements Evictor {
                     evictedRecordCounter++;
                 }
             }
-            if (evictedRecordCounter >= removalSize) {
+            if (evictedRecordCounter >= sizeToEvict) {
                 break;
             }
         }
@@ -150,11 +173,14 @@ public class EvictorImpl implements Evictor {
         return index < 0 ? 0 : index;
     }
 
-    @Override
-    public int findRemovalSize(RecordStore recordStore) {
+    protected int calculateRemovalSize(RecordStore recordStore) {
+        int size = recordStore.size();
+        if (size == 0) {
+            return 0;
+        }
+
         MapConfig mapConfig = recordStore.getMapContainer().getMapConfig();
         int maxSize = mapConfig.getMaxSizeConfig().getSize();
-        int currentPartitionSize = recordStore.size();
 
         int removalSize;
         final MaxSizeConfig.MaxSizePolicy maxSizePolicy = mapConfig.getMaxSizeConfig().getMaxSizePolicy();
@@ -163,8 +189,8 @@ public class EvictorImpl implements Evictor {
             case PER_PARTITION:
                 int targetSizePerPartition = Double.valueOf(maxSize
                         * ((ONE_HUNDRED_PERCENT - evictionPercentage) / (1D * ONE_HUNDRED_PERCENT))).intValue();
-                int diffFromTargetSize = currentPartitionSize - targetSizePerPartition;
-                int prunedSize = currentPartitionSize * evictionPercentage / ONE_HUNDRED_PERCENT + 1;
+                int diffFromTargetSize = size - targetSizePerPartition;
+                int prunedSize = size * evictionPercentage / ONE_HUNDRED_PERCENT + 1;
                 removalSize = Math.max(diffFromTargetSize, prunedSize);
                 break;
             case PER_NODE:
@@ -174,8 +200,8 @@ public class EvictorImpl implements Evictor {
                         * memberCount / mapServiceContext.getNodeEngine().getPartitionService().getPartitionCount());
                 targetSizePerPartition = Double.valueOf(maxPartitionSize
                         * ((ONE_HUNDRED_PERCENT - evictionPercentage) / (1D * ONE_HUNDRED_PERCENT))).intValue();
-                diffFromTargetSize = currentPartitionSize - targetSizePerPartition;
-                prunedSize = currentPartitionSize * evictionPercentage / ONE_HUNDRED_PERCENT + 1;
+                diffFromTargetSize = size - targetSizePerPartition;
+                prunedSize = size * evictionPercentage / ONE_HUNDRED_PERCENT + 1;
                 removalSize = Math.max(diffFromTargetSize, prunedSize);
                 break;
             case USED_HEAP_PERCENTAGE:
@@ -183,7 +209,7 @@ public class EvictorImpl implements Evictor {
             case FREE_HEAP_PERCENTAGE:
             case FREE_HEAP_SIZE:
                 // if we have an evictable size, be sure to evict at least one entry in worst case.
-                removalSize = Math.max(currentPartitionSize * evictionPercentage / ONE_HUNDRED_PERCENT, 1);
+                removalSize = Math.max(size * evictionPercentage / ONE_HUNDRED_PERCENT, 1);
                 break;
             default:
                 throw new IllegalArgumentException("Max size policy is not defined [" + maxSizePolicy + "]");
@@ -203,5 +229,207 @@ public class EvictorImpl implements Evictor {
         }
         return value;
     }
+
+
+    @Override
+    public boolean isReachedMaxSize(RecordStore recordStore) {
+        if (recordStore.size() == 0) {
+            return false;
+        }
+        String mapName = recordStore.getName();
+        int partitionId = recordStore.getPartitionId();
+
+        MapContainer mapContainer = recordStore.getMapContainer();
+        MaxSizeConfig maxSizeConfig = mapContainer.getMapConfig().getMaxSizeConfig();
+        MaxSizeConfig.MaxSizePolicy maxSizePolicy = maxSizeConfig.getMaxSizePolicy();
+
+        boolean result;
+        switch (maxSizePolicy) {
+            case PER_NODE:
+                return checkPerNodeEviction(mapName, maxSizeConfig);
+            case PER_PARTITION:
+                result = checkPerPartitionEviction(mapName, maxSizeConfig, partitionId);
+                break;
+            case USED_HEAP_PERCENTAGE:
+                result = checkHeapPercentageEviction(mapName, maxSizeConfig);
+                break;
+            case USED_HEAP_SIZE:
+                result = checkHeapSizeEviction(mapName, maxSizeConfig);
+                break;
+            case FREE_HEAP_PERCENTAGE:
+                result = checkFreeHeapPercentageEviction(maxSizeConfig);
+                break;
+            case FREE_HEAP_SIZE:
+                result = checkFreeHeapSizeEviction(maxSizeConfig);
+                break;
+            default:
+                throw new IllegalArgumentException("Not an appropriate max size policy [" + maxSizePolicy + ']');
+        }
+        return result;
+    }
+
+
+    protected boolean checkPerNodeEviction(String mapName, MaxSizeConfig maxSizeConfig) {
+        long nodeTotalSize = 0;
+
+        final double maxSize = getApproximateMaxSize(maxSizeConfig.getSize());
+        final List<Integer> partitionIds = findPartitionIds();
+        for (int partitionId : partitionIds) {
+            final PartitionContainer container = mapServiceContext.getPartitionContainer(partitionId);
+            if (container == null) {
+                continue;
+            }
+            nodeTotalSize += getRecordStoreSize(mapName, container);
+            if (nodeTotalSize >= maxSize) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    protected boolean checkPerPartitionEviction(String mapName, MaxSizeConfig maxSizeConfig, int partitionId) {
+        final double maxSize = getApproximateMaxSize(maxSizeConfig.getSize());
+        final PartitionContainer container = mapServiceContext.getPartitionContainer(partitionId);
+        if (container == null) {
+            return false;
+        }
+        final int size = getRecordStoreSize(mapName, container);
+        return size >= maxSize;
+    }
+
+    protected boolean checkHeapSizeEviction(String mapName, MaxSizeConfig maxSizeConfig) {
+        final long usedHeapSize = getUsedHeapSize(mapName);
+        if (usedHeapSize == -1L) {
+            return false;
+        }
+        final double maxSize = getApproximateMaxSize(maxSizeConfig.getSize());
+        return maxSize < (1D * usedHeapSize / ONE_MEGABYTE);
+    }
+
+    protected boolean checkFreeHeapSizeEviction(MaxSizeConfig maxSizeConfig) {
+        final long currentFreeHeapSize = getAvailableMemory();
+        final double minFreeHeapSize = getApproximateMaxSize(maxSizeConfig.getSize());
+        return minFreeHeapSize > (1D * currentFreeHeapSize / ONE_MEGABYTE);
+    }
+
+    protected boolean checkHeapPercentageEviction(String mapName, MaxSizeConfig maxSizeConfig) {
+        final long usedHeapSize = getUsedHeapSize(mapName);
+        if (usedHeapSize == -1L) {
+            return false;
+        }
+        final double maxSize = getApproximateMaxSize(maxSizeConfig.getSize());
+        final long total = getTotalMemory();
+        return maxSize < (1D * ONE_HUNDRED_PERCENT * usedHeapSize / total);
+    }
+
+    protected boolean checkFreeHeapPercentageEviction(MaxSizeConfig maxSizeConfig) {
+        final long currentFreeHeapSize = getAvailableMemory();
+        final double freeHeapPercentage = getApproximateMaxSize(maxSizeConfig.getSize());
+        final long total = getTotalMemory();
+        return freeHeapPercentage > (1D * ONE_HUNDRED_PERCENT * currentFreeHeapSize / total);
+    }
+
+    protected long getTotalMemory() {
+        return memoryInfoAccessor.getTotalMemory();
+    }
+
+    protected long getFreeMemory() {
+        return memoryInfoAccessor.getFreeMemory();
+    }
+
+    protected long getMaxMemory() {
+        return memoryInfoAccessor.getMaxMemory();
+    }
+
+    protected long getAvailableMemory() {
+        final long totalMemory = getTotalMemory();
+        final long freeMemory = getFreeMemory();
+        final long maxMemory = getMaxMemory();
+        return freeMemory + (maxMemory - totalMemory);
+    }
+
+    protected long getUsedHeapSize(String mapName) {
+        long heapCost = 0L;
+        final List<Integer> partitionIds = findPartitionIds();
+        for (int partitionId : partitionIds) {
+            final PartitionContainer container = mapServiceContext.getPartitionContainer(partitionId);
+            if (container == null) {
+                continue;
+            }
+            heapCost += getRecordStoreHeapCost(mapName, container);
+        }
+
+        MapContainer mapContainer = mapServiceContext.getMapContainer(mapName);
+        heapCost += mapContainer.getNearCacheSizeEstimator().getSize();
+        return heapCost;
+    }
+
+    protected int getRecordStoreSize(String mapName, PartitionContainer partitionContainer) {
+        final RecordStore existingRecordStore = partitionContainer.getExistingRecordStore(mapName);
+        if (existingRecordStore == null) {
+            return 0;
+        }
+        return existingRecordStore.size();
+    }
+
+    protected long getRecordStoreHeapCost(String mapName, PartitionContainer partitionContainer) {
+        final RecordStore existingRecordStore = partitionContainer.getExistingRecordStore(mapName);
+        if (existingRecordStore == null) {
+            return 0L;
+        }
+        return existingRecordStore.getHeapCost();
+    }
+
+    /**
+     * used when deciding evictable or not.
+     */
+    protected static double getApproximateMaxSize(int maxSizeFromConfig) {
+        // because not to exceed the max size much we start eviction early.
+        // so decrease the max size with ratio .95 below
+        return 1D * maxSizeFromConfig * EVICTION_START_THRESHOLD_PERCENTAGE / ONE_HUNDRED_PERCENT;
+    }
+
+    /**
+     * Get max size setting form config for given policy
+     *
+     * @return max size or -1 if policy is different or not set
+     */
+    public static double getApproximateMaxSize(MaxSizeConfig maxSizeConfig, MaxSizeConfig.MaxSizePolicy policy) {
+        if (maxSizeConfig.getMaxSizePolicy() == policy) {
+            return getApproximateMaxSize(maxSizeConfig.getSize());
+        }
+        return -1D;
+    }
+
+    protected List<Integer> findPartitionIds() {
+        final NodeEngine nodeEngine = mapServiceContext.getNodeEngine();
+        final InternalPartitionService partitionService = nodeEngine.getPartitionService();
+        final int partitionCount = partitionService.getPartitionCount();
+        List<Integer> partitionIds = null;
+        for (int partitionId = 0; partitionId < partitionCount; partitionId++) {
+            if (isOwnerOrBackup(partitionId)) {
+                if (partitionIds == null) {
+                    partitionIds = new ArrayList<Integer>();
+                }
+                partitionIds.add(partitionId);
+            }
+        }
+        return partitionIds == null ? Collections.<Integer>emptyList() : partitionIds;
+    }
+
+
+    protected boolean isOwnerOrBackup(int partitionId) {
+        final NodeEngine nodeEngine = mapServiceContext.getNodeEngine();
+        final InternalPartitionService partitionService = nodeEngine.getPartitionService();
+        final InternalPartition partition = partitionService.getPartition(partitionId, false);
+        final Address thisAddress = nodeEngine.getThisAddress();
+        return partition.isOwnerOrBackup(thisAddress);
+    }
+
+    // only used when testing.
+    public void setMemoryInfoAccessor(MemoryInfoAccessor memoryInfoAccessor) {
+        this.memoryInfoAccessor = memoryInfoAccessor;
+    }
+
 
 }
